% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoading.R
\name{coordinatesFromVariants}
\alias{coordinatesFromVariants}
\title{Convert From Coordinates to Variants}
\usage{
coordinatesFromVariants(variants, chromCol, offset = 0)
}
\description{
Convert From Coordinates to Variants
}
