% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVAnalysis.R
\name{generatePositionHeatmap}
\alias{generatePositionHeatmap}
\title{Generates a vector of NV Occurrences}
\usage{
generatePositionHeatmap(variants)
}
\arguments{
\item{variants}{A data frame containing the variants with ALT and REF columns
in the format that exists in VCFs. Typical usage employs loading this data
via linkVariantsWithMetadata.}
}
\description{
Takes a data frame of known nucleotides variants and counts their occurrences
on different positions. Groups occurrences based on the variant allele.
}
\examples{
heatmap <- PhenoGenRLib::generatePositionHeatmap(huntingtonsVariants)
# Use View(heatmap) to look at the dataset.

}
\seealso{
\code{\link[=linkVariantsWithMetadata]{linkVariantsWithMetadata()}}
}
