% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huntingtonsVariants}
\alias{huntingtonsVariants}
\title{Demo Huntingtons Patient Variants Linked with Metadata}
\format{
An object of class \code{data.frame} with 312 rows and 62 columns.
}
\source{
NCBI BioProject PRJNA299309
}
\usage{
huntingtonsVariants
}
\description{
FASTQ files obtained from BioProject with accession PRJNA299309 was loaded
into Bowtie2, for alignment, and Freebayes for variant calling. Produced
Variant Call Format (VCF) files were loaded with the linking function
built-in to the library.
}
\details{
This is a data frame produced from reading multiple VCFs and linking their
associated metadata. On top of the the standard tags typically seen in a VCF
this file also contains the filename itself, an additional chromosome tag
with human readable numbers indicating the chromosome of origin for the
sample, and a \code{dummy_pheno}, representing whether or not this patient has
phenotype A, or B.
}
\seealso{
\code{\link[=linkVariantsWithMetadata]{linkVariantsWithMetadata()}}
}
\keyword{datasets}
