% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huntingtonsVariantsHeatmap}
\alias{huntingtonsVariantsHeatmap}
\title{Demo Huntingtons Patient Variants Heatmap}
\format{
An object of class \code{data.frame} with 6946 rows and 8 columns.
}
\source{
NCBI BioProject PRJNA299309
}
\usage{
huntingtonsVariantsHeatmap
}
\description{
A Dataframe containing 8 columns, two columns representing the reference,
and the alternate alleles of A, T, C and G. Heatmap generated from the
data set \code{huntingtonsVariants} by use of the heatmap function.
}
\seealso{
\code{\link[=generatePositionHeatmap]{generatePositionHeatmap()}}
}
\keyword{datasets}
