% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoading.R
\name{linkVariantsWithMetadata}
\alias{linkVariantsWithMetadata}
\title{Link Existing VCF files with Metadata}
\usage{
linkVariantsWithMetadata(metadataFile, vcfDir, vcfColName)
}
\arguments{
\item{metadataFile}{A path to the metadata file in the CSV format. This
file must at the very least, contain one column which lists the various
VCF's filenames. For each filename, associated metadata should appear
on the same row.}

\item{vcfDir}{A path to the directory containing the VCFs. The previously
mentioned filenames will be appended in a smart manner to from the full path.}

\item{vcfColName}{The name of the column containing the VCF files.}
}
\value{
Returns a data.frame like object where each row is a variant and
the associated metadata is attached column-wise to the right.
}
\description{
This function's purpose is to join a series of VCFs together and
associate sample, or other metadata with each of the variants inside
each of the previously mentioned VCFs. To do this, one column in the
metadata CSV file must contain the file names of each VCF, and on the
same row, contain the associated metadata with that VCF. Then,
it looks inside \code{vcfDir} for the VCF filenames as they appear in the
CSV.
}
\examples{
variants <- PhenoGenRLib::linkVariantsWithMetadata(
  metadata = "inst/extdata/huntingtons_datasheet_shortened.csv",
  vcfDir = "inst/extdata/",
  vcfColName = "vcfs"
)
View(variants)

}
