% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mappedHuntingtonsVariants}
\alias{mappedHuntingtonsVariants}
\title{Demo Huntingtons Patient Variants with Metadata and Mapped to Known RSIDs}
\format{
An object of class \code{list} of length 2.
}
\source{
NCBI BioProject PRJNA299309
}
\usage{
mappedHuntingtonsVariants
}
\description{
Similar to \code{huntingtonVariants}. This dataset is derived from the previous
by additionally calling the RSID mapping function on \code{huntingtonVariants}.
}
\details{
This dataset is a list of two different data frames, one of which under the
the name \code{mappedHuntingtonVariants$nvs} is largely the same with the
\code{huntingtonVariants} data frame, however, contains extra information on the
associated RSIDs. The second data frame is a list of all RSIDs pulled and
contains more information for each of the RSIDs. The latter can be found
under the name \code{mappedHuntingtonVariants$rsids}.
}
\seealso{
\code{\link[=mapRsidsForVariants]{mapRsidsForVariants()}}
}
\keyword{datasets}
