% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MappedVariants}
\alias{MappedVariants}
\title{Demo Variants Loaded and Linked with Metadata}
\format{
An object of class \code{data.frame} with 489 rows and 64 columns.
}
\source{
BioProject PRJNA299309
}
\usage{
MappedVariants
}
\description{
Variants from BioProject with accession PRJNA299309 was loaded
via use of the library function \code{linkVariantsWithMetadata} and
\code{mapRsidsForVariants}.
}
\keyword{datasets}
