% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoading.R
\name{mapRsidsForVariants}
\alias{mapRsidsForVariants}
\title{Maps RefSeq IDs to the variants}
\usage{
mapRsidsForVariants(
  chromCol,
  variants,
  offset = 0,
  hostGenVersion = 38,
  batchSize = 100,
  progress = NULL
)
}
\arguments{
\item{chromCol}{The column containing the name of the chromosome. It may be
beneficial to use the metadata file to add this as often times, the VCF's chromosome
is a ID rather than just a number.}

\item{variants}{The variant data.frame to map rsIDs for.}

\item{offset}{VCFs report positions relative to the reference sequence. Of course,
if this reference sequence is not the entire chromosome, the reported coordinate will
not match with any known rsIDs. Therefore, an offset may need to be introduced. E.g.,
if the reference sequence started at 3074681 (ex. NC_000004.12:3074681-3243960), we may
may choose the offset to be 3074680. Default is 0 (no offset).}

\item{hostGenVersion}{The Host genome version. The two supported options are 38 (GRCh38),
and 37 (GRCh37). Default is 38.}

\item{batchSize}{The number coordinates to send to Ensembl servers at once. Smaller batches
may reduce chances of timing out. Default is 100.}

\item{progress}{A function to call each iteration to track progress. The
function should accept two numeric parameters, one representing the current
iteration, and the second representing the total number of iterations.}
}
\value{
Returns variant data.frame containing all pre-existing information (including metadata and
other built-in VCF annotations), as well as the associated rsIDs, chromosomal and known chromosomal
start and end positions. Also returns a data.frame of all the rsIDs and associated information such
allele frequencies in the databases. The two values are returned in a data.frame where \code{nvs} is the
accessor for the variants, and \code{rsids} is the accessor for the rsID information.
}
\description{
This function takes a data.frame like object containing the variant
information and attached metadata. It then searches the ensembl database
to find rsIDs for the well-known variants that appear. The rsIDs are then
added column-wise to each of the variants in the original data.frame. This
updated data.frame is what is returned.
}
\examples{

mappedVariants <- mapRsidsForVariants(
chromCol = "chromosome",
variants = huntingtonsVariants,
offset = 3074680
)
# Use View(mappedVariants$nvs) to see the variants and their mappings
# Use View(mappedVariants$rsids) to see just the RSID information
}
