% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RsIDs}
\alias{RsIDs}
\title{Demo RsIDs}
\format{
An object of class \code{data.frame} with 217 rows and 8 columns.
}
\source{
Ensembl https://doi.org/10.1093/nar/gkac958
}
\usage{
RsIDs
}
\description{
Reference sequence identifiers fetched from Ensembl.
\code{mapRsidsForVariants}.
}
\keyword{datasets}
