% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPhenoGenRLibShiny.R
\name{runPhenoGenRLib}
\alias{runPhenoGenRLib}
\title{Launch Shiny App For PhenoGenRLib}
\usage{
runPhenoGenRLib()
}
\value{
Opens a Shiny application but does not return anything.
}
\description{
Launches Shiny app for this package. The application allows for loading
VCFs, linking said VCFs to their case metadata, resolving RSIDs based on
chromosome and position, and running associative case-control based on
Ensembl population. The app also allows for easier visualization of variance
distribution based on different nucleobases.
}
\examples{
\dontrun{
runPhenoGenRLib()
}


}
