% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnmappedVariants}
\alias{UnmappedVariants}
\title{Demo Variants Loaded}
\format{
An object of class \code{data.frame} with 312 rows and 61 columns.
}
\source{
BioProject PRJNA299309
}
\usage{
UnmappedVariants
}
\description{
Variants from BioProject with accession PRJNA299309 was loaded
via use of the library function \code{linkVariantsWithMetadata}.
}
\keyword{datasets}
