% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variants}
\alias{variants}
\title{Demo Variants Loaded and Linked with Metadata}
\format{
An object of class \code{data.frame} with 312 rows and 61 columns.
}
\source{
BioProject PRJNA299309
}
\usage{
variants
}
\description{
Variants from BioProject with accession PRJNA299309 was loaded
via use of the library function \code{mapRsidsForVariants}.
}
\keyword{datasets}
