% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{visVariantDistribution}
\alias{visVariantDistribution}
\title{Visualize Variant Distribution}
\usage{
visVariantDistribution(variants)
}
\arguments{
\item{variants}{The data.frame containing all the variant
information.}
}
\description{
Sometimes, it may be beneficial to know approximately which
genomic regions have the most detected variants relative to
reference sequence. This function generates a simple figure
depicting the positions on the horizontal axis and the number
of variants on the vertical axis.
}
\examples{
plot <- visVariantDistribution(MappedVariants)
View(plot)

}
