% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{visVariantHeatMap}
\alias{visVariantHeatMap}
\title{Visualize Variant Heat Map}
\usage{
visVariantHeatMap(heatmap, nucbases)
}
\arguments{
\item{heatmap}{The data frame containing the variant distribution of the
various nucleotides at each positions.}

\item{nucbases}{The nucleotides to graph. Can be any combination of
"A", "T", "C", or "G".}
}
\description{
Sometimes, it may be beneficial to know approximately which
genomic regions have the most detected variants relative to
reference sequence. This function generates a simple figure
depicting the positions on the horizontal axis and the number
of variants on the vertical axis.
}
\examples{

plot <- visVariantHeatMap(huntingtonsVariantsHeatmap, nucbases = c("A", "T"))
# This is a ggplot2 plot

}
