% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVAnalysis.R
\name{generatePositionHeatmap}
\alias{generatePositionHeatmap}
\title{Generates a vector of NV Occurrences}
\usage{
generatePositionHeatmap(variants)
}
\description{
Takes a data frame of known nucleotides variants and counts their occurrences
on different positions. Groups ocurrences based on the variant allele.
}
