% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoading.R
\name{mapRsidsForVariantPositions}
\alias{mapRsidsForVariantPositions}
\title{Maps RSIDs based on Variant Positions}
\usage{
mapRsidsForVariantPositions(coordinates, hostGenVersion = 38)
}
\description{
Helper function for fetching RSIDs based off set of specific coordinates
}
