% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVAnalysis.R
\name{multipleAssociationTests}
\alias{multipleAssociationTests}
\title{Performs Multiple Association Tests on Arrays of Matrices}
\usage{
multipleAssociationTests(mxns, useChi = FALSE, groupName = "Untitled")
}
\description{
Performs Multiple Association Tests on Arrays of Matrices
}
